Flow migration command shouldn't blow up if a single flow fails
